requirements.txt
astroid==1.3.8
logilab-common==1.0.2
Pillow==2.9.0
pylint==1.4.4
pytesseract==0.1.6
six==1.9.0
wheel==0.24.0
